Add Close and unexport unnecessary things.
Add a close method in block files, and add a check in Directory that sees if the
file is still accessible from its original location (closing it if not), so we
can actually clean up disk space.  Also, unexport exported objects from preadlib
that we don't use in the main stenographer binary.