Switch from an external to a local SSTable implementation.
The original sstable implementation we pulled from an external leveldb
implementation failed... it appears as though it was optimized to return the
last value for a key, where we want ALL values for a key.  This might be doable
with leveldb... maybe it only allows one value per key at the higher levels?
But it 100% doesn't work for us... we store MANY values for each key.

The code is relatively untested, and I really would rather have not written it,
but I couldn't find another leveldb implementation that exposed the internals.
Also, the implementation we used initially was really complex... lots of knobs
that I just had to guess at.  This one exposes a very clean interface that does
just what we want and no more.

Needs much testing, etc, but appears to work MUCH better than the old code did.
Huzzah!