Refactor query parsing into its own library.
Initiallly, we passed queries around as strings, then each index parsed the
query independently and handled it.  Now, the query is parsed once into a nice
structure, then that structure is applied to all indexes to get the set of
positions we need.

Also, changed base.Int64Slice to base.Positions, since it seems to detail what
we're actually using it for much more correctly.