Update index file format to be nicer to leveldb.
LevelDB really expects that its underlying tables will have one value per key...
we've abused that in the past, but we're switching to that model.  Where we used
to have:

  Key: "<indextype><indexvalue>"  Value: "<blockfileposition>"

We now have:

  Key: "<indextype><indexvalue><blockfileposition>"  Value: ""

This means we can remove the hacks in indexfile for finding the start of
iteration by looking for the previous key.  It also plays nicer with more strict
versions of leveldb which may assert that keys ascend without repeating.