Fix index version write, added watchdogs.
Fixed the index write of the version number, which I'd mistakenly put BEFORE I
actually created the sstable to write to :-/.  I should build things before
committing.

Also, added watchdog threads to the major, important threads in stenotype, so if
any thread blocks for a substantial amount of time we blow up the world.