Fix error checking when opening files.
We used to have Errno() take a bool, and return an error if it was
false.  Now, we have Errno() take an int, and return an error if that
int is < 0.  However, we still appear to have a few cases floating
around where it is passed a bool.  Since a bool is either 0 or 1, this
means we never return an error. :(

Fixed in two places, and updated comment.