Modify cert usage; allow users to manage certs.
Steno's use of certificates for user/group verification initially was designed a
little wonky:

 1) the server cert was used to sign the client cert
 2) certs (both client and server) were recreated from scratch on each run of
    steno

These two issues made it really hard to do anything more complex with certs,
like having a single client cert that works across multiple servers, or using an
organization's central CA for cert management.

This CL changes that:

  1) Certificate creation is now done at install time (stenokeys.sh) with the
     'openssl' command-line tool, not inside Go code.  Thus, keys are no longer
     rewritten on startup.
  2) There is now a CA cert in /etc/stenographer/certs/ca_cert.pem.  This CA
     cert is used to sign both the client and server cert.
     * Users can now overwrite it with their org's CA cert to centrally manage
       certificates.
     * Users can generate multiple client or server certs from a single CA cert
       and manually distribute them.

This is somewhat backwards-incompatible with old installs of steno, which don't
have a ca_cert.  Thus, when install.sh is run now, if it sees that ca_cert is
missing, it'll wipe all certs and reinstall all certs from scratch.

Note, this should also fix Issue #124, where NSS TLS is more stringent with
certs and doesn't allow a CA cert to also act as a server cert.