blockfile: Fix overwriting 'blockData' before packet is read.
When `stenoread` is used to query all packets in blocks,
struct 'allPacketsIter' is used.  The entire block
is read into 'allPacketsIter.blockData' and packets
to be sent to `stenoread` all refer to this byte slice.
Then, after one block is read, the blockfile module will
immediately load the next block.  Since the 'blockData'
is statically allocated, the loading of next block will
overwrite the data of yet to be sent packets, causing
corruption when `stenoread` tries to parse the packets
using `tcpdump`.

To fix it, this commit forces `allPacketsIter` function
to always `make([]byte, 1 << 20)` a new byte slice for
the next block, so that previous block data is always
retained until all packets in the block are sent to
`stenoread`.

Reported-by: Reid Price <reid@awakenetworks.com>
Signed-off-by: Alex Wang <alex@awakenetworks.com>