Update steno.conf
Adding in the default configurations into the stenographer's configuration file will enhance the self-documentation of the tool. This will make it easier to understand the behaviors of the tool without referencing the GitHub document.
The default values were obtained from:
https://github.com/google/stenographer/blob/master/INSTALL.md

To quote the the INSTALL.md document:

"MaxDirectoryFiles: The maximum number of packet/index files to create before cleaning old ones up. Defaults to 30K files..."
"Note that DiskFreePercentage is optional... it defaults to 10%."