Adds note about acceptable relative time values.
Took me a while to debug my app (https://github.com/rocknsm/docket) because I was parsing duration strings with Python libraries and rendering them to seconds as a common output. So `5m` would work, but when I converted it to `300.0s` it failed because "s" isn't supported, and the "." made the lexer assume it was an ip address. This note hopes to document that requirement.